backend/requirements.txt
# API server
fastapi
uvicorn[standard]
pydantic-settings
prometheus_client

# HTTP client (overpass, osrm)
httpx

# AI visibility filter
groq

# Geospatial math (haversine, bearing, project_endpoint)
geopy
geographiclib

# Ray casting + UTM projection (visibility pipeline)
shapely
pyproj

# Environment variables
python-dotenv

# Supabase (auth + persistent profiles)
supabase>=2.0.0
psycopg2-binary

# Testing
pytest
pytest-asyncio
